V0.8.0 stacked

+++ plugin/core/src/main/java/group/aelysium/rustyconnector/core/lib/algorithm/QuickSort.java
@@ -1,6 +1,6 @@
 package group.aelysium.rustyconnector.core.lib.algorithm;
 
-import group.aelysium.rustyconnector.api.velocity.load_balancing.ISortable;
+import group.aelysium.rustyconnector.toolkit.velocity.load_balancing.ISortable;
 
 import java.util.List;
 
@@ -12,22 +12,13 @@ public class QuickSort {
     }
 
     static int partition(List<? extends ISortable> array, int low, int high) {
-        // pivot
-        int pivot = array.get(high).sortIndex();
-  
-        // Index of smaller element and
-        // indicates the right position
-        // of pivot found so far
+        double pivot = array.get(high).sortIndex();
+
         int i = (low - 1);
   
         for (int j = low; j <= high - 1; j++) {
-  
-            // If current element is smaller
-            // than the pivot
+
             if (array.get(j).sortIndex() < pivot) {
-  
-                // Increment index of
-                // smaller element
                 i++;
                 swap(array, i, j);
             }
@@ -48,6 +39,8 @@ public class QuickSort {
     /**
      * The main quicksort implementation.
      * Assumes that start is 0 and the final index is equal to size.
+     * This method will mutate the array that is passed.
+     * Sorts from least to greatest.
      * @param array The array to sort.
      */
     public static void sort(List<? extends ISortable> array) {
