+++ plugin/core/src/main/java/group/aelysium/rustyconnector/core/lib/algorithm/SortableInteger.java
@@ -1,6 +1,6 @@
 package group.aelysium.rustyconnector.core.lib.algorithm;
 
-import group.aelysium.rustyconnector.api.velocity.load_balancing.ISortable;
+import group.aelysium.rustyconnector.toolkit.velocity.load_balancing.ISortable;
 
 public class SortableInteger implements ISortable {
     int value;
@@ -10,7 +10,7 @@ public class SortableInteger implements ISortable {
     }
 
     @Override
-    public int sortIndex() {
+    public double sortIndex() {
         return this.value;
     }
